Bug fixes
Fixed several bugs in the code, that would misname the component, making the program unable to issue commands.

If a rorze component is now connected (eTRB0) it is renamed to (oTRB0) which is the required command prefix. HOWEVER this is not true for the simulation. There is a way prettier fix than this, will do this later.