Updated Menu
The menu will now only show IPs at first. Then in the background the software will start communicating with all connected components and ask for an acknowledgement. If the acknowledgement comes from a component, the software will ask for Serial number and the component Type. The goal is to have this updated live in the main menu screen, even before a connection to a component is established by the user.