Added a status window
A status window can now pop up in all of the menus, that has a certain duration. This can prevent the user from connecting to a component with multiple threads, which would result in an instant error. Instead, a popup window will ask the user to wait until the first TCP socket is closed.